AAE-24331 move ProcessVariableSpecification to query-rest
